Split out WebSocketError into file
This commit moves the WebSocketError type into its own file to help keep
the module organized. This is the first step towards being able to move
these types to the rest api crates.
